+++ src/Build5Nines.SharpVector/IVectorDatabaseExtensions.cs
@@ -1,3 +1,5 @@
+using System.Text;
+
 namespace Build5Nines.SharpVector;
 
 public static class IVectorDatabaseExtensions
@@ -7,7 +9,7 @@ public static class IVectorDatabaseExtensions
     {
         using (var stream = new FileStream(filePath, FileMode.Create, FileAccess.Write))
         {
-            await vectorDatabase.SerializeToJsonStreamAsync(stream);
+            await vectorDatabase.SerializeToBinaryStreamAsync(stream);
         }
     }
 
@@ -16,7 +18,7 @@ public static class IVectorDatabaseExtensions
     {
         using (var stream = new FileStream(filePath, FileMode.Create, FileAccess.Write))
         {
-            vectorDatabase.SerializeToJsonStream(stream);
+            vectorDatabase.SerializeToBinaryStream(stream);
         }
     }
 
@@ -25,16 +27,16 @@ public static class IVectorDatabaseExtensions
     {
         using (var stream = new FileStream(filePath, FileMode.Open, FileAccess.Read))
         {
-            await vectorDatabase.DeserializeFromJsonStreamAsync(stream);
+            await vectorDatabase.DeserializeFromBinaryStreamAsync(stream);
         }
     }
 
-    public static void LoadFromFile<TId, TMetadata, TDocument>(IVectorDatabase<TId, TMetadata, TDocument> vectorDatabase, string filePath)
+    public static void LoadFromFile<TId, TMetadata, TDocument>(this IVectorDatabase<TId, TMetadata, TDocument> vectorDatabase, string filePath)
         where TId : notnull
     {
         using (var stream = new FileStream(filePath, FileMode.Open, FileAccess.Read))
         {
-            vectorDatabase.DeserializeFromJsonStream(stream);
+            vectorDatabase.DeserializeFromBinaryStream(stream);
         }
     }
 }