+++ src/SharpVectorTest/VectorDatabaseTests.cs
@@ -78,6 +78,94 @@ public class VectorDatabaseTests
     }
 
     [TestMethod]
+    public void BasicMemoryVectorDatabase_SaveLoadAsync_01()
+    {
+        var vdb = new BasicMemoryVectorDatabase();
+        
+        // // Load Vector Database with some sample text
+        vdb.AddText("The Lion King is a 1994 Disney animated film about a young lion cub named Simba who is the heir to the throne of an African savanna.", "[some metadata here]");
+        
+        var results = vdb.Search("Lion King");
+
+        Assert.AreEqual(1, results.Texts.Count());
+        Assert.IsTrue(results.Texts.First().Text.Contains("Lion King"));
+        Assert.AreEqual("[some metadata here]", results.Texts.First().Metadata);
+        Assert.AreEqual(0.3396831452846527, results.Texts.First().VectorComparison);
+
+        var filename = "basicmemoryvectordatabase_saveload_01.b59vdb";
+        vdb.SaveToFileAsync(filename).Wait();
+
+
+        var newvdb = new BasicMemoryVectorDatabase();
+        newvdb.LoadFromFileAsync(filename).Wait();
+        results = newvdb.Search("Lion King");
+
+        Assert.AreEqual(1, results.Texts.Count());
+        Assert.IsTrue(results.Texts.First().Text.Contains("Lion King"));
+        Assert.AreEqual("[some metadata here]", results.Texts.First().Metadata);
+        Assert.AreEqual(0.3396831452846527, results.Texts.First().VectorComparison);
+    }
+
+    [TestMethod]
+    public void BasicMemoryVectorDatabase_SaveLoad_01()
+    {
+        var vdb = new BasicMemoryVectorDatabase();
+        
+        // // Load Vector Database with some sample text
+        vdb.AddText("The Lion King is a 1994 Disney animated film about a young lion cub named Simba who is the heir to the throne of an African savanna.", "[some metadata here]");
+        
+        var results = vdb.Search("Lion King");
+
+        Assert.AreEqual(1, results.Texts.Count());
+        Assert.IsTrue(results.Texts.First().Text.Contains("Lion King"));
+        Assert.AreEqual("[some metadata here]", results.Texts.First().Metadata);
+        Assert.AreEqual(0.3396831452846527, results.Texts.First().VectorComparison);
+
+        var filename = "basicmemoryvectordatabase_saveload_01.b59vdb";
+        vdb.SaveToFile(filename);
+
+
+        var newvdb = new BasicMemoryVectorDatabase();
+        newvdb.LoadFromFile(filename);
+        results = newvdb.Search("Lion King");
+
+        Assert.AreEqual(1, results.Texts.Count());
+        Assert.IsTrue(results.Texts.First().Text.Contains("Lion King"));
+        Assert.AreEqual("[some metadata here]", results.Texts.First().Metadata);
+        Assert.AreEqual(0.3396831452846527, results.Texts.First().VectorComparison);
+    }
+
+    [TestMethod]
+    public async Task BasicMemoryVectorDatabase_SaveLoadBinaryStreamAsync_01()
+    {
+        var vdb = new BasicMemoryVectorDatabase();
+        
+        // // Load Vector Database with some sample text
+        vdb.AddText("The Lion King is a 1994 Disney animated film about a young lion cub named Simba who is the heir to the throne of an African savanna.", "[some metadata here]");
+        
+        var results = vdb.Search("Lion King");
+
+        Assert.AreEqual(1, results.Texts.Count());
+        Assert.IsTrue(results.Texts.First().Text.Contains("Lion King"));
+        Assert.AreEqual("[some metadata here]", results.Texts.First().Metadata);
+        Assert.AreEqual(0.3396831452846527, results.Texts.First().VectorComparison);
+
+        var stream = new MemoryStream();
+        await vdb.SerializeToBinaryStreamAsync(stream);
+
+        stream.Position = 0; // Reset the stream position to the beginning
+        
+        var newvdb = new BasicMemoryVectorDatabase();
+        await newvdb.DeserializeFromBinaryStreamAsync(stream);
+        results = newvdb.Search("Lion King");
+
+        Assert.AreEqual(1, results.Texts.Count());
+        Assert.IsTrue(results.Texts.First().Text.Contains("Lion King"));
+        Assert.AreEqual("[some metadata here]", results.Texts.First().Metadata);
+        Assert.AreEqual(0.3396831452846527, results.Texts.First().VectorComparison);
+    }
+
+    [TestMethod]
     public void SimpleTest_01()
     {
         var vdb = new MemoryVectorDatabase<double>();
@@ -476,12 +564,12 @@ public class VectorDatabaseTests
         Assert.AreEqual(0.3396831452846527, firstResult.Texts.First().VectorComparison);
 
         var stream = new MemoryStream();
-        databaseOne.SerializeToJsonStream(stream);
+        databaseOne.SerializeToBinaryStream(stream);
         
         stream.Position = 0; // move to beginning of stream
 
         var databaseTwo = new MemoryVectorDatabase<double>();
-        databaseTwo.DeserializeFromJsonStream(stream);
+        databaseTwo.DeserializeFromBinaryStream(stream);
 
         var secondResult = await databaseTwo.SearchAsync("Lion King", pageCount: 5);
         Assert.AreEqual(5, secondResult.Texts.Count());
@@ -645,7 +733,7 @@ public class VectorDatabaseTests
         var id = vdb.AddText("The Lion King is a 1994 Disney animated film about a young lion cub named Simba who is the heir to the throne of an African savanna.", "{ value: \"JSON Metadata Value\" }");
         
         var stream = new MemoryStream();
-        await vdb.SerializeToJsonStreamAsync(stream);
+        await vdb.SerializeToBinaryStreamAsync(stream);
         stream.Position = 0;
 
         var databaseInfo = await DatabaseFile.LoadDatabaseInfoAsync(stream);
@@ -664,7 +752,7 @@ public class VectorDatabaseTests
         var id = vdb.AddText("The Lion King is a 1994 Disney animated film about a young lion cub named Simba who is the heir to the throne of an African savanna.", "{ value: \"JSON Metadata Value\" }");
         
         var stream = new MemoryStream();
-        await vdb.SerializeToJsonStreamAsync(stream);
+        await vdb.SerializeToBinaryStreamAsync(stream);
         stream.Position = 0;
 
         vdb = await DatabaseFile.Load<MemoryVectorDatabase<string>, string>(stream);
@@ -691,7 +779,7 @@ public class VectorDatabaseTests
         var id = vdb.AddText("The Lion King is a 1994 Disney animated film about a young lion cub named Simba who is the heir to the throne of an African savanna.", "{ value: \"JSON Metadata Value\" }");
         
         var stream = new MemoryStream();
-        await vdb.SerializeToJsonStreamAsync(stream);
+        await vdb.SerializeToBinaryStreamAsync(stream);
         stream.Position = 0;
 
         vdb = await DatabaseFile.Load<string>(stream);